Added superuser sercurity for updating services and general login security for viewing profiles

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Service, Animal, Category
 from .forms import ServiceForm
 # Create your views here.
@@ -33,8 +34,12 @@ def service_detail(request, service_id):
     return render(request, 'products/service_detail.html', context)
 
 
+@login_required
 def add_service(request):
     """ Add a new service to the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Hey, you are not authorised to be down here!')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = ServiceForm(request.POST, request.FILES)
         if form.is_valid():
@@ -53,8 +58,13 @@ def add_service(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_service(request, service_id):
     """ Edit an existing service on the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Hey, you are not authorised to be down here!')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Service, pk=service_id)
     if request.method == 'POST':
         form = ServiceForm(request.POST, request.FILES, instance=service)
@@ -77,8 +87,13 @@ def edit_service(request, service_id):
         return render(request, template, context)
 
 
+@login_required
 def delete_service(request, service_id):
     """ Delete service from site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Hey, you are not authorised to be down here!')
+        return redirect(reverse('home'))
+        
     service = get_object_or_404(Service, pk=service_id)
     service.delete()
     messages.success(request, 'Service gone forever...')
